[ios, osx] Reorganized geometry classes
Renamed MGLGeometryCollection to MGLShapeCollection for consistency with MGLShape, which is the base of all the concrete geometry classes. Moved MGLShapeCollection into its own files and the MGLFeature classes into MGLFeature.h/.mm. Grouped classes in the Foundation and SDK groups by theme.

Fixed an issue in which shape collections were not being initialized with any data.