[ios, osx] Corrected MGLFeature naming; documented MGLFeature
Renamed featureTag to identifier because the GeoJSON specification allows feature IDs to be strings, whereas “tag” in Cocoa typically refers to a number-typed identifier. Also renamed featureAttributes and -objectForKey: to be consistent with more modern Cocoa APIs that have open-ended key-value storage. Added documentation on the various new classes, including a detailed description of how feature attributes work. Added the new classes to the jazzy table of contents.